Move class CubeMesh
