+++ src/Kononov/Common/Scene/CubeMesh.cpp
