+++ src/Kononov/Common/Scene/CubeMesh.hpp
