+++ src/Kononov/Task1/MainWindow.cpp
@@ -3,8 +3,8 @@
 #include <QOpenGLFunctions>
 #include <QScreen>
 
-#include <CubeMesh.hpp>
 #include <Resources.hpp>
+#include <Scene/CubeMesh.hpp>
 #include <Shader.hpp>
 #include <Vertex.hpp>
